backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1